Chewie has now a REST API

+++ src/main/java/com/chewie/repositories/UserRepository.java
@@ -1,13 +1,13 @@
 package com.chewie.repositories;
-import java.util.Optional;
 
 import com.chewie.domain.User;
-
 import org.springframework.data.repository.CrudRepository;
 
+import java.util.Optional;
 
 
-public interface UserRepository extends CrudRepository<User, Integer> {
+
+public interface UserRepository extends CrudRepository<User, Long> {
     /**
      * Save a given User
      * 
@@ -22,5 +22,5 @@ public interface UserRepository extends CrudRepository<User, Integer> {
      * @param id must not be null
      * @return the User with the given id or empty
      */
-    Optional<User> findById(Integer id);
+    Optional<User> findById(Long id);
 }
